+++ src/main/java/com/shrhang/create_food_filling/content/listener/EatingListener.java
@@ -3,6 +3,7 @@ package com.shrhang.create_food_filling.content.listener;
 import com.shrhang.create_food_filling.api.event.AnimalFeedingEvent;
 import com.shrhang.create_food_filling.api.event.FoxEatingEvent;
 import com.shrhang.create_food_filling.api.event.PandaEatingEvent;
+import com.shrhang.create_food_filling.api.event.VillagerEatingEvent;
 import net.neoforged.neoforge.common.NeoForge;
 import net.neoforged.neoforge.event.entity.living.LivingEntityUseItemEvent;
 
@@ -13,8 +14,9 @@ public class EatingListener {
     public static void init() {
         NeoForge.EVENT_BUS.addListener(EatingListener::onEat);
         NeoForge.EVENT_BUS.addListener(EatingListener::onAnimalFeeding);
-        NeoForge.EVENT_BUS.addListener(EatingListener::onPandaEating);
         NeoForge.EVENT_BUS.addListener(EatingListener::onFoxEating);
+        NeoForge.EVENT_BUS.addListener(EatingListener::onPandaEating);
+        NeoForge.EVENT_BUS.addListener(EatingListener::onVillagerEating);
     }
 
     public static void onEat(final LivingEntityUseItemEvent.Finish event) {
@@ -25,11 +27,15 @@ public class EatingListener {
         tryToApplyEffect(event.getItemStack(), event.getEntity());
     }
 
+    public static void onFoxEating(final FoxEatingEvent event) {
+        tryToApplyEffect(event.getItemStack(), event.getEntity());
+    }
+
     public static void onPandaEating(final PandaEatingEvent event) {
         tryToApplyEffect(event.getItemStack(), event.getEntity());
     }
 
-    public static void onFoxEating(final FoxEatingEvent event) {
+    public static void onVillagerEating(final VillagerEatingEvent event) {
         tryToApplyEffect(event.getItemStack(), event.getEntity());
     }
 }