+++ src/main/java/com/shrhang/create_food_filling/content/util/ApplyEffectUtil.java
@@ -1,7 +1,6 @@
 package com.shrhang.create_food_filling.content.util;
 
 import com.shrhang.create_food_filling.Config;
-import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
 
@@ -10,7 +9,7 @@ import static net.minecraft.core.component.DataComponents.POTION_CONTENTS;
 
 public class ApplyEffectUtil {
     public static boolean preFilter(ItemStack stack, LivingEntity entity) {
-        if (!Config.COMMON.prefilterEvent.get()) return true;
+        if (!Config.SERVER.isPreFilterEvent.get()) return true;
 
         return Config.COMMON.isEatingApplyEffects.get()
                 && !entity.level().isClientSide