+++ src/main/java/com/shrhang/create_food_filling/mixin/PandaMixin.java
@@ -1,5 +1,6 @@
 package com.shrhang.create_food_filling.mixin;
 
+import com.shrhang.create_food_filling.Config;
 import com.shrhang.create_food_filling.api.event.PandaEatingEvent;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.animal.Panda;
@@ -16,12 +17,12 @@ import static com.shrhang.create_food_filling.content.util.ApplyEffectUtil.preFi
 public abstract class PandaMixin {
     @Inject(method = "handleEating", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/animal/Panda;setItemSlot(Lnet/minecraft/world/entity/EquipmentSlot;Lnet/minecraft/world/item/ItemStack;)V"))
     private void create_food_filling$onPandaEating(CallbackInfo ci) {
+        if (!Config.SERVER.enablePandaFoodEffects.get()) return;
         Panda panda = (Panda) (Object) this;
         ItemStack itemStack = panda.getItemBySlot(EquipmentSlot.MAINHAND);
         if (!preFilter(itemStack, panda)) return;
         if (!itemStack.isEmpty()) {
-            PandaEatingEvent event = new PandaEatingEvent(panda, itemStack);
-            NeoForge.EVENT_BUS.post(event);
+            NeoForge.EVENT_BUS.post(new PandaEatingEvent(panda, itemStack));
         }
     }
 }